+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Pending changes
 
 - [#83](https://github.com/bumble-tech/appyx/issues/83) – **Breaking change**: `RoutingSource` renamed to `NavModel`. All subclasses, fields, package names, etc., any mentions of the word follow suit.
-
+- [#91](https://github.com/bumble-tech/appyx/pull/91) – **Fixed**: Spotlight next and previous operations crash fix 
 
 ## 1.0-alpha04